refactor: move EntityNameSanitizer to utils layer
