--- src/data/usecases/sanitize-name/sanitize-entity-name-usecase.ts
@@ -1,7 +1,0 @@
-import { SanitizeEntityName } from "../../../domain/usecases/sanitize-entity-name";
-
-export class DbSanitizeEntityName implements SanitizeEntityName {
-  sanitize(name: string): string {
-    return name.replace(/-/g, " ").toLocaleLowerCase();
-  }
-}
