+++ src/presentation/controllers/provinces/find-province-controller.spec.ts
@@ -1,7 +1,7 @@
-import { SanitizeEntityName } from "../../../domain/usecases/sanitize-entity-name";
 import { ProvinceModel } from "../../../domain/models";
 import { CheckProvinceByName } from "../../../domain/usecases/check-province-by-name";
 import { FindProvince } from "../../../domain/usecases/find-province";
+import { SanitizeEntityName } from "../../protocols/sanitize-entity-name";
 import {
   FindProvinceController,
   LoadProvinceController,
