+++ src/presentation/protocols/sanitize-entity-name.ts
