+++ src/utils/sanitize-entity/sanitize-entity-name.spec.ts
@@ -1,8 +1,8 @@
-import { DbSanitizeEntityName } from "./sanitize-entity-name-usecase";
+import { EntityNameSanitizer } from "./sanitize-entity-name";
 
 describe("DbSanitizeEntityName", () => {
   it("santizes the passed name", () => {
-    const sut = new DbSanitizeEntityName();
+    const sut = new EntityNameSanitizer();
     expect(sut.sanitize("british-columbia")).toBe("british columbia");
     expect(sut.sanitize("British-columbia")).toBe("british columbia");
     expect(sut.sanitize("British-COlumBIa")).toBe("british columbia");
